refactor: split service route adapters by domain (#522)
